Add desktop_webview_auth plugin and refresh token logic for desktop platform
In the given diff, code for authentication on the desktop platform was added. This change includes the implementation of the `desktop_webview_auth` plugin in the CMake, cc and Dart files. The version of the Flutter application has also been updated.

The pubspec.yaml file was modified to include new dependencies (desktop_webview_auth and firebase_core_desktop) and the Google Sign-In process for desktop platforms was also implemented. The modified files now handle sign-in processes for both mobile and desktop platforms separately, improving the overall user experience across all platforms.